refactor: update import paths and remove unused home page
